--- src/app/system/components/button-wyszukaj/button-wyszukaj.component.html
@@ -1,1 +1,0 @@
-<button mat-raised-button>Wyszukaj</button>
