--- src/app/system/components/button-wyszukaj/button-wyszukaj.component.less
@@ -1,5 +1,0 @@
-button {
-    width: 250px;
-    height: 60px;
-    margin-right: 32px;
-}
