--- src/app/system/components/button-wyszukaj/button-wyszukaj.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'srs-button-wyszukaj',
-  templateUrl: './button-wyszukaj.component.html',
-  styleUrls: ['./button-wyszukaj.component.less']
-})
-export class ButtonWyszukajComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
