--- src/app/system/components/combobox-pojemnosc/combobox-pojemnosc.component.html
@@ -1,29 +1,0 @@
-<h2>Pojemonosc</h2>
-
-<div id="all">
-    <div id="od">
-        <p>Od</p>
-
-        <mat-form-field appearance="fill">
-            <mat-label>Liczba</mat-label>
-            <mat-select>
-                <mat-option *ngFor="let liczby of zbior" [value]="liczby.value">
-                    {{ liczby.viewValue }}
-                </mat-option>
-            </mat-select>
-        </mat-form-field>
-    </div>
-
-    <div id="do">
-        <p>Do</p>
-
-        <mat-form-field appearance="fill">
-            <mat-label>Liczba</mat-label>
-            <mat-select>
-                <mat-option *ngFor="let liczby of zbior" [value]="liczby.value">
-                    {{ liczby.viewValue }}
-                </mat-option>
-            </mat-select>
-        </mat-form-field>
-    </div>
-</div>
