--- src/app/system/components/combobox-pojemnosc/combobox-pojemnosc.component.less
@@ -1,29 +1,0 @@
-mat-form-field {
-    margin-left: 32px;
-    margin-right: 30px;
-    width: 100px;
-}
-
-p {
-    margin-left: 32px;
-    margin-right: 32px;
-}
-
-h2 {
-    margin-left: 32px;
-    margin-right: 32px;
-}
-
-#od {
-    display: inline-block;
-    width: 130;
-}
-
-#do {
-    display: inline-block;
-    width: 130;
-}
-
-#all {
-    flex-direction: row;
-}
