--- src/app/system/components/combobox-typ-sali/combobox-typ-sali.component.html
@@ -1,8 +1,0 @@
-<mat-form-field appearance="fill">
-    <mat-label>Typ sali</mat-label>
-    <mat-select>
-        <mat-option *ngFor="let typ of typy" [value]="typ.value">
-            {{ typ.viewValue }}
-        </mat-option>
-    </mat-select>
-</mat-form-field>
