--- src/app/system/components/combobox-typ-sali/combobox-typ-sali.component.less
@@ -1,4 +1,0 @@
-mat-form-field {
-    margin-left: 32px;
-    margin-right: 32px;
-}
