--- src/app/system/components/combobox-wydzial/combobox-wydzial.component.less
@@ -1,4 +1,0 @@
-mat-form-field {
-    margin-left: 32px;
-    margin-right: 32px;
-}
