--- src/app/system/components/comboboxdaty/comboboxdaty.component.html
@@ -1,39 +1,0 @@
-<div id="all">
-    <div id="data">
-        <h2>Data</h2>
-        <mat-form-field appearance="fill">
-            <mat-label>Choose a date</mat-label>
-            <input matInput [matDatepicker]="picker" />
-            <mat-datepicker-toggle
-                matSuffix
-                [for]="picker"
-            ></mat-datepicker-toggle>
-            <mat-datepicker #picker></mat-datepicker>
-        </mat-form-field>
-    </div>
-    <div id="od">
-        <p>Od</p>
-
-        <mat-form-field appearance="fill">
-            <mat-label>Godzina</mat-label>
-            <mat-select>
-                <mat-option *ngFor="let liczby of zbior" [value]="liczby.value">
-                    {{ liczby.viewValue }}
-                </mat-option>
-            </mat-select>
-        </mat-form-field>
-    </div>
-
-    <div id="do">
-        <p>Do</p>
-
-        <mat-form-field appearance="fill">
-            <mat-label>Godzina</mat-label>
-            <mat-select>
-                <mat-option *ngFor="let liczby of zbior" [value]="liczby.value">
-                    {{ liczby.viewValue }}
-                </mat-option>
-            </mat-select>
-        </mat-form-field>
-    </div>
-</div>
