--- src/app/system/components/comboboxdaty/comboboxdaty.component.less
@@ -1,34 +1,0 @@
-mat-form-field {
-    margin-left: 32px;
-    margin-right: 32px;
-    width: 90%;
-}
-
-p {
-    margin-left: 32px;
-    margin-right: 32px;
-}
-
-h2 {
-    margin-left: 32px;
-    margin-right: 32px;
-}
-
-#data {
-    display: inline-block;
-    width: 15%;
-}
-
-#od {
-    display: inline-block;
-    width: 6%;
-}
-
-#do {
-    display: inline-block;
-    width: 6%;
-}
-
-#all {
-    flex-direction: row;
-}
