--- src/app/system/components/nrsali/nrsali.component.html
@@ -1,7 +1,0 @@
-<h2>Nr sali</h2>
-<form class="example-form">
-    <mat-form-field appearance="fill">
-        <mat-label>Nr sali</mat-label>
-        <input matInput placeholder="" value="" />
-    </mat-form-field>
-</form>
