--- src/app/system/components/nrsali/nrsali.component.less
@@ -1,10 +1,0 @@
-h2 {
-    margin-left: 32px;
-    margin-right: 32px;
-}
-
-mat-form-field {
-    margin-left: 32px;
-    margin-right: 32px;
-    width: 260px;
-}
