--- src/app/system/components/nrsali/nrsali.component.ts
@@ -1,13 +1,0 @@
-import { Component, OnInit } from '@angular/core'
-
-@Component({
-    selector: 'srs-nrsali',
-    templateUrl: './nrsali.component.html',
-    styleUrls: ['./nrsali.component.less'],
-})
-export class NrsaliComponent implements OnInit {
-    value = 'Clear me'
-    constructor() {}
-
-    ngOnInit(): void {}
-}
