+++ src/app/system/components/parametry-wyszukiwania/parametry-wyszukiwania.component.html
@@ -1,0 +1,133 @@
+<main>
+    <section>
+        <mat-form-field appearance="fill">
+            <mat-label>Wydział</mat-label>
+            <mat-select>
+                <mat-option
+                    *ngFor="let wydzialy of wydzialyy"
+                    [value]="wydzialy.value"
+                >
+                    {{ wydzialy.viewValue }}
+                </mat-option>
+            </mat-select>
+        </mat-form-field>
+
+        <mat-form-field appearance="fill">
+            <mat-label>Typ sali</mat-label>
+            <mat-select>
+                <mat-option *ngFor="let typ of typy" [value]="typ.value">
+                    {{ typ.viewValue }}
+                </mat-option>
+            </mat-select>
+        </mat-form-field>
+    </section>
+    <h2>Pojemonosc</h2>
+    <section>
+        <div id="od">
+            <p>Od</p>
+            <mat-form-field class="comboboxLiczba" appearance="fill">
+                <mat-label>Liczba</mat-label>
+                <mat-select>
+                    <mat-option
+                        *ngFor="let liczby of zbior"
+                        [value]="liczby.value"
+                    >
+                        {{ liczby.viewValue }}
+                    </mat-option>
+                </mat-select>
+            </mat-form-field>
+        </div>
+        <div id="do">
+            <p>Do</p>
+            <mat-form-field class="comboboxLiczba" appearance="fill">
+                <mat-label>Liczba</mat-label>
+                <mat-select>
+                    <mat-option
+                        *ngFor="let liczby of zbior"
+                        [value]="liczby.value"
+                    >
+                        {{ liczby.viewValue }}
+                    </mat-option>
+                </mat-select>
+            </mat-form-field>
+        </div>
+    </section>
+    <h2>Nr sali</h2>
+    <section>
+        <form class="example-form">
+            <mat-form-field class="nrSali" appearance="fill">
+                <mat-label>Nr sali</mat-label>
+                <input matInput placeholder="" value="" />
+            </mat-form-field>
+        </form>
+    </section>
+    <h2>Data</h2>
+    <section>
+        <div id="data">
+            <p>data</p>
+            <mat-form-field class="dataField" appearance="fill">
+                <mat-label>Choose a date</mat-label>
+                <input matInput [matDatepicker]="picker" />
+                <mat-datepicker-toggle
+                    matSuffix
+                    [for]="picker"
+                ></mat-datepicker-toggle>
+                <mat-datepicker #picker></mat-datepicker>
+            </mat-form-field>
+        </div>
+        <div id="odData">
+            <p>Od</p>
+            <mat-form-field class="comboboxLiczbaPojemnosc" appearance="fill">
+                <mat-label>Godzina</mat-label>
+                <mat-select>
+                    <mat-option
+                        *ngFor="let liczby of zbior"
+                        [value]="liczby.value"
+                    >
+                        {{ liczby.viewValue }}
+                    </mat-option>
+                </mat-select>
+            </mat-form-field>
+        </div>
+        <div id="doData">
+            <p>Do</p>
+            <mat-form-field class="comboboxLiczbaPojemnosc" appearance="fill">
+                <mat-label>Godzina</mat-label>
+                <mat-select>
+                    <mat-option
+                        *ngFor="let liczby of zbior"
+                        [value]="liczby.value"
+                    >
+                        {{ liczby.viewValue }}
+                    </mat-option>
+                </mat-select>
+            </mat-form-field>
+        </div>
+    </section>
+    <section>
+        <div id="checkBox">
+            <h2>Wyposazenie</h2>
+            <section class="checkBox">
+                <mat-checkbox class="checkBoxRzutnik">Rzutnik!</mat-checkbox>
+            </section>
+
+            <section class="checkBox">
+                <mat-checkbox class="checkBoxKlimatyzacjian"
+                    >Klimatyzacjia!</mat-checkbox
+                >
+            </section>
+            <section class="checkBox">
+                <mat-checkbox class="checkBoxKomputer">Komputer!</mat-checkbox>
+            </section>
+            <section class="checkBox">
+                <mat-checkbox class="checkBoxTablica">Tablica!</mat-checkbox>
+            </section>
+            <section class="checkBox">
+                <mat-checkbox class="checkBoxInne">Inne!</mat-checkbox>
+            </section>
+        </div>
+        <div id="buttonWyszukaj">
+            <button class="buttonWyszukaj" mat-raised-button>Wyszukaj</button>
+        </div>
+    </section>
+</main>
