+++ src/app/system/components/parametry-wyszukiwania/parametry-wyszukiwania.component.less
@@ -1,0 +1,49 @@
+mat-form-field,
+h2,
+p {
+    margin-left: 32px;
+    margin-right: 32px;
+}
+
+.comboboxLiczba {
+    width: 100px;
+}
+
+.comboboxLiczbaPojemnosc {
+    width: 100px;
+}
+
+.checkBox {
+    margin-left: 32px;
+}
+
+.nrSali {
+    width: 265px;
+}
+
+.dataField {
+    width: 265px;
+}
+
+.buttonWyszukaj {
+    width: 265px;
+    height: 50px;
+}
+
+#checkBox {
+    width: 360px;
+}
+
+main {
+    width: 800px;
+    display: flex;
+    flex-direction: column;
+}
+section {
+    display: flex;
+    align-items: center;
+}
+
+#buttonWyszukaj {
+    align-items: center;
+}
