+++ src/app/system/components/tabela-wyszukiwania-sale/tabela-wyszukiwania-sale.component.html
