+++ src/app/system/components/tabela-wyszukiwania-sale/tabela-wyszukiwania-sale.component.less
@@ -1,0 +1,21 @@
+table {
+    margin-bottom: 32px;
+    margin-right: 32px;
+    width: 500px;
+}
+
+p {
+    margin-bottom: 32px;
+    margin-right: 32px;
+    width: 500px;
+}
+
+h2 {
+    margin-bottom: 32px;
+    margin-right: 32px;
+    width: 500px;
+}
+
+srs-tabela-wyszukiwania-sale {
+    width: 500px;
+}
