--- src/app/system/components/wynik-wyszukiwanie-sal/wynik-wyszukiwanie-sal.component.less
@@ -1,9 +1,0 @@
-table {
-    margin-bottom: 32px;
-    margin-right: 32px;
-}
-
-p {
-    margin-bottom: 32px;
-    margin-right: 32px;
-}
