--- src/app/system/components/wyposazenie/wyposazenie.component.html
@@ -1,20 +1,0 @@
-<h2>Wyposazenie</h2>
-<section class="checkBoxRzutnik">
-    <mat-checkbox class="checkBoxRzutnik">Rzutnik!</mat-checkbox>
-</section>
-
-<section class="checkBoxKlimatyzacjia">
-    <mat-checkbox class="checkBoxKlimatyzacjian">Klimatyzacjia!</mat-checkbox>
-</section>
-
-<section class="checkBoxKomputer">
-    <mat-checkbox class="checkBoxKomputer">Komputer!</mat-checkbox>
-</section>
-
-<section class="checkBoxTablica">
-    <mat-checkbox class="checkBoxTablica">Tablica!</mat-checkbox>
-</section>
-
-<section class="checkBoxInne">
-    <mat-checkbox class="checkBoxInne">Inne!</mat-checkbox>
-</section>
