--- src/app/system/components/wyposazenie/wyposazenie.component.less
@@ -1,8 +1,0 @@
-h2 {
-    margin-left: 32px;
-    margin-right: 32px;
-}
-section {
-    margin-left: 32px;
-    margin-right: 32px;
-}
