--- src/app/system/components/wyposazenie/wyposazenie.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'srs-wyposazenie',
-  templateUrl: './wyposazenie.component.html',
-  styleUrls: ['./wyposazenie.component.less']
-})
-export class WyposazenieComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
