+++ src/app/system/pages/wyszukiwarka-sal/wyszukiwarka-sal.component.html
@@ -1,9 +1,9 @@
-<h1>Wyszukiwanie sali</h1>
-<section><srs-wynik-wyszukiwanie-sal></srs-wynik-wyszukiwanie-sal>></section>
-<srs-combobox-wydzial></srs-combobox-wydzial>
-<srs-combobox-typ-sali></srs-combobox-typ-sali>
-<srs-combobox-pojemnosc></srs-combobox-pojemnosc>
-<srs-nrsali></srs-nrsali>
-<srs-comboboxdaty></srs-comboboxdaty>
-<srs-wyposazenie></srs-wyposazenie>
-<srs-button-wyszukaj></srs-button-wyszukaj>
+<!-- total_hours_wasted_here: 8-->
+<main>
+    <h1>Wyszukiwanie sali</h1>
+    <section>
+        <srs-parametry-wyszukiwania></srs-parametry-wyszukiwania>
+
+        <srs-tabela-wyszukiwania-sale></srs-tabela-wyszukiwania-sale>
+    </section>
+</main>
