+++ src/app/system/pages/wyszukiwarka-sal/wyszukiwarka-sal.component.less
@@ -2,9 +2,14 @@ h1 {
     padding-left: 32px;
     padding-top: 32px;
     font-size: 2rem;
+    width: 500px;
 }
+main {
+    width: 600px;
+    display: flex;
+    flex-direction: column;
+}
+
 section {
-    padding: 32px;
-    position: absolute;
-    left: 1000px;
+    display: flex;
 }
